requirements.txt
selenium==4.40.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
google-generativeai==0.6.0
pymupdf==1.24.9
streamlit==1.53.1
requests==2.32.3
